+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/CDU/A320_Neo_CDU_DirectToPage.js
@@ -34,8 +34,10 @@ class CDUDirectToPage {
             });
         };
         let i = 0;
-        while (i < mcdu.flightPlanManager.getWaypointsCount() && i + wptsListIndex < mcdu.flightPlanManager.getWaypointsCount() && i < iMax) {
-            const waypoint = mcdu.flightPlanManager.getWaypoint(i + wptsListIndex);
+        wptsListIndex = Math.max(wptsListIndex, mcdu.flightPlanManager.getActiveWaypointIndex());
+        const totalWaypointsCount = mcdu.flightPlanManager.getWaypointsCount() + mcdu.flightPlanManager.getArrivalWaypointsCount() + mcdu.flightPlanManager.getApproachWaypointsCount();
+        while (i < totalWaypointsCount && i + wptsListIndex < totalWaypointsCount && i < iMax) {
+            const waypoint = mcdu.flightPlanManager.getWaypoint(i + wptsListIndex, NaN, true);
             if (waypoint) {
                 waypointsCell[i] = "←" + waypoint.ident + "[color]blue";
                 if (waypointsCell[i]) {
@@ -85,7 +87,7 @@ class CDUDirectToPage {
         ]);
         mcdu.onUp = () => {
             wptsListIndex++;
-            wptsListIndex = Math.min(wptsListIndex, mcdu.flightPlanManager.getWaypointsCount() - 4);
+            wptsListIndex = Math.min(wptsListIndex, totalWaypointsCount - 5);
             CDUDirectToPage.ShowPage(mcdu, directWaypoint, wptsListIndex);
         };
         mcdu.onDown = () => {